+++ src/api/endpoints/EndpointsContentReference.ts
@@ -1,8 +1,8 @@
-import { sendRequest } from "./sendRequest";
+import { sendRequest } from "../requests/sendRequest";
 import { itRexVars } from "../../constants/Constants";
 import { ApiUrls } from "../../constants/ApiUrls";
 import { loggerFactory } from "../../../logger/LoggerConfig";
-import { ResponseParser } from "./ResponseParser";
+import { ResponseParser } from "../responses/ResponseParser";
 import { IEndpointsContentReference } from "../endpoints_interfaces/IEndpointsContentReference";
 import { IContent } from "../../types/IContent";
